refactor: move ModelLHSReader.js -> generate/model-lhs-reader.js
